move git to admin command. Revert ping
